+++ portlets/opensocial-portlet/docroot/WEB-INF/service/com/liferay/opensocial/service/OAuthTokenLocalServiceUtil.java
@@ -234,7 +234,7 @@ public class OAuthTokenLocalServiceUtil {
 		long userId, long gadgetId, java.lang.String serviceName,
 		long moduleId, java.lang.String accessToken,
 		java.lang.String tokenName, java.lang.String tokenSecret,
-		java.lang.String sessionHandle, int expiration)
+		java.lang.String sessionHandle, long expiration)
 		throws com.liferay.portal.kernel.exception.PortalException,
 			com.liferay.portal.kernel.exception.SystemException {
 		return getService()
